# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/{shorten}/count": {
            "get": {
                "description": "shorten url redirection count.",
                "tags": [
                    "api"
                ],
                "summary": "shorten url redirection count.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shorten Url Code",
                        "name": "shorten",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
}